@@ -1,4 +1,4 @@
-var dict = [];
+const dict: Record<number, string> = {};
 
 dict[0x3007] = "líng,xīng"; /* 〇: https://zh.m.wiktionary.org/wiki/%E3%80%87 */
 dict[0x3400] = "qiū"; /* 㐀 */
@@ -6886,7 +6886,7 @@ dict[0x545e] = "shī"; /* 呞 */
 dict[0x545f] = "juǎn"; /* 呟 */
 dict[0x5460] = "pěn"; /* 呠 */
 dict[0x5461] = "wěn,mǐn"; /* 呡 */
-dict[0x5462] = "ní,ne"; /* 呢 */
+dict[0x5462] = "ne,ní"; /* 呢 */
 dict[0x5463] = "móu"; /* 呣 */
 dict[0x5464] = "líng"; /* 呤 */
 dict[0x5465] = "rán"; /* 呥 */
@@ -15788,7 +15788,7 @@ dict[0x773c] = "yǎn"; /* 眼 */
 dict[0x773d] = "mò"; /* 眽 */
 dict[0x773e] = "zhòng"; /* 眾 */
 dict[0x773f] = "mò"; /* 眿 */
-dict[0x7740] = "zhuó,zháo,zhāo,zhe"; /* 着 */
+dict[0x7740] = "zháo,zhuó,zhāo,zhe"; /* 着 */
 dict[0x7741] = "zhēng"; /* 睁 */
 dict[0x7742] = "méi"; /* 睂 */
 dict[0x7743] = "suō"; /* 睃 */
@@ -19108,7 +19108,7 @@ dict[0x8439] = "biān"; /* 萹 */
 dict[0x843a] = "mào"; /* 萺 */
 dict[0x843b] = "ān"; /* 萻 */
 dict[0x843c] = "è"; /* 萼 */
-dict[0x843d] = "là,luò,lào"; /* 落 */
+dict[0x843d] = "luò,là,lào"; /* 落 */
 dict[0x843e] = "yíng"; /* 萾 */
 dict[0x843f] = "kuò"; /* 萿 */
 dict[0x8440] = "kuò"; /* 葀 */
@@ -26150,7 +26150,7 @@ dict[0x20007] = "qiě"; /*  */
 dict[0x20008] = "qiū"; /*  */
 dict[0x20009] = "qiū"; /* 	 */
 dict[0x2000a] = "cāo"; /* */
-dict[0x2000c] = "qiū"; /*  */
+dict[0x2000c] = "qiū"; /* */
 dict[0x2000d] = "shì"; /*  */
 dict[0x20013] = "sī"; /*  */
 dict[0x20014] = "jué"; /*  */
@@ -28767,10 +28767,10 @@ dict[0x21ffa] = "nǎo"; /* Ὼ */
 dict[0x21ffc] = "xùn"; /* ῼ */
 dict[0x21ffe] = "jú"; /* ῾ */
 dict[0x21fff] = "huò"; /* ῿ */
-dict[0x22001] = "yì"; /*   */
-dict[0x2200a] = "xī"; /*   */
-dict[0x2200b] = "sè"; /* ​ */
-dict[0x2200c] = "jiǎo"; /* ‌ */
+dict[0x22001] = "yì"; /* */
+dict[0x2200a] = "xī"; /* */
+dict[0x2200b] = "sè"; /* */
+dict[0x2200c] = "jiǎo"; /* */
 dict[0x2200d] = "yōng"; /* ‍ */
 dict[0x22013] = "jù"; /* – */
 dict[0x22015] = "shī"; /* ― */
@@ -30112,7 +30112,7 @@ dict[0x22ff1] = "sàn"; /* ⿱ */
 dict[0x22ff2] = "qiāo"; /* ⿲ */
 dict[0x22ff3] = "luàn"; /* ⿳ */
 dict[0x22ff8] = "cè"; /* ⿸ */
-dict[0x23000] = "léi"; /* 　 */
+dict[0x23000] = "léi"; /* */
 dict[0x23001] = "zhǎn"; /* 、 */
 dict[0x23002] = "lǐ"; /* 。 */
 dict[0x23003] = "lián"; /* 〃 */
@@ -41246,4 +41246,4 @@ dict[0x2a590] = "yào,yà"; /* ꖐ */
 dict[0x2a595] = "huī"; /* ꖕ */
 dict[0x2a599] = "hōu"; /* ꖙ */
 
-module.exports = dict;
+export default dict;
