@@ -1,6 +1,7 @@
 // 有些通假字，单字本身没有这个读音，但是通假字可以有。
 // 如：阿家阿翁 [["ā"], ["gū"], ["ā"], ["wēng"]]，家通姑。
-module.exports = {
+
+const phrases_dict: Record<string,string[][]> = {
 "阿Ｑ": [["ā"], ["Ｑ"]],
 "阿爸": [["ā"], ["bà"]],
 "阿鼻": [["ā"], ["bí"]],
@@ -1564,7 +1565,7 @@ module.exports = {
 "倡优": [["chàng"], ["yōu"]],
 "唱和": [["chàng"], ["hè"]],
 "唱头": [["chàng"], ["tóu"]],
-"抄查": [["chāo"], ["zhā"]],
+"抄查": [["chāo"], ["chá"]],
 "抄家": [["chāo"], ["jiā"]],
 "钞票": [["chāo"], ["piào"]],
 "抄没": [["chāo"], ["méi"]],
@@ -7287,7 +7288,6 @@ module.exports = {
 "碣石": [["jié"], ["shí"]],
 "竭尽": [["jié"], ["jìn"]],
 "羯羊": [["jié"], ["yáng"]],
-"她们": [["tā"], ["men"]],
 "姐夫": [["jiě"], ["fū"]],
 "解表": [["jiě"], ["biǎo"]],
 "解差": [["jiè"], ["chāi"]],
@@ -7466,6 +7466,7 @@ module.exports = {
 "进给": [["jìn"], ["jǐ"]],
 "进港": [["jìn"], ["gǎng"]],
 "进化": [["jìn"], ["huà"]],
+"进行": [["jìn"], ["xíng"]],
 "进见": [["jìn"], ["jiàn"]],
 "进剿": [["jìn"], ["jiǎo"]],
 "进食": [["jìn"], ["shí"]],
@@ -9728,6 +9729,7 @@ module.exports = {
 "泥子": [["nì"], ["zǐ"]],
 "霓裳": [["ní"], ["cháng"]],
 "你好": [["nǐ"], ["hǎo"]],
+"你们": [["nǐ"], ["men"]],
 "昵称": [["nì"], ["chēng"]],
 "拟合": [["nǐ"], ["hé"]],
 "逆差": [["nì"], ["chà"]],
@@ -10281,7 +10283,7 @@ module.exports = {
 "凭据": [["píng"], ["jù"]],
 "凭空": [["píng"], ["kōng"]],
 "凭信": [["píng"], ["xìn"]],
-"苹果": [["pín"], ["guǒ"]],
+"苹果": [["píng"], ["guǒ"]],
 "屏除": [["bǐng"], ["chú"]],
 "屏蔽": [["píng"], ["bì"]],
 "屏风": [["píng"], ["fēng"]],
@@ -12630,8 +12632,11 @@ module.exports = {
 "锁骨": [["suǒ"], ["gǔ"]],
 "锁上": [["suǒ"], ["shàng"]],
 "锁钥": [["suǒ"], ["yuè"]],
-"他处": [["tā"], ["chǔ"]],
+"他处": [["tā"], ["chù"]],
 "他家": [["tā"], ["jiā"]],
+"他们": [["tā"], ["men"]],
+"她们": [["tā"], ["men"]],
+"它们": [["tā"], ["men"]],
 "趿拉": [["tā"], ["lā"]],
 "塌落": [["tā"], ["luò"]],
 "塌台": [["tā"], ["tái"]],
@@ -13717,6 +13722,7 @@ module.exports = {
 "蜗旋": [["wō"], ["xuán"]],
 "我见": [["wǒ"], ["jiàn"]],
 "我家": [["wǒ"], ["jiā"]],
+"我们": [["wǒ"], ["men"]],
 "沃地": [["wò"], ["dì"]],
 "卧车": [["wò"], ["chē"]],
 "卧倒": [["wò"], ["dǎo"]],
@@ -17410,7 +17416,7 @@ module.exports = {
 "埃塞俄比亚": [["āi"], ["sāi"], ["é"], ["bǐ"], ["yà"]],
 "挨冻受饿": [["ái"], ["dòng"], ["shòu"], ["è"]],
 "挨风缉缝": [["āi"], ["fēng"], ["jī"], ["fèng"]],
-"挨家挨户": [["āi"], ["ji"], ["āi"], ["hù"]],
+"挨家挨户": [["āi"], ["jiā"], ["āi"], ["hù"]],
 "挨肩并足": [["āi"], ["jiān"], ["bìng"], ["zú"]],
 "挨肩擦背": [["āi"], ["jiān"], ["cā"], ["bèi"]],
 "挨肩擦膀": [["āi"], ["jiān"], ["cā"], ["bǎng"]],
@@ -18863,7 +18869,7 @@ module.exports = {
 "不合时宜": [["bù"], ["hé"], ["shí"], ["yí"]],
 "不哼不哈": [["bù"], ["hēng"], ["bù"], ["hā"]],
 "不护细行": [["bù"], ["hù"], ["xì"], ["xíng"]],
-"不怀好意": [["bù"], ["huái"], ["hào"], ["yì"]],
+"不怀好意": [["bù"], ["huái"], ["hǎo"], ["yì"]],
 "不欢而散": [["bù"], ["huān"], ["ér"], ["sàn"]],
 "不慌不忙": [["bù"], ["huāng"], ["bù"], ["máng"]],
 "不遑枚举": [["bù"], ["huáng"], ["méi"], ["jǔ"]],
@@ -20817,7 +20823,7 @@ module.exports = {
 "打交道": [["dǎ"], ["jiāo"], ["dào"]],
 "打街骂巷": [["dǎ"], ["jiē"], ["mà"], ["xiàng"]],
 "打瞌睡": [["dǎ"], ["kē"], ["shuì"]],
-"打开天窗说亮话": [["dǎ"], ["kāi"], ["tiān"], ["chuāng"], ["liàng"], ["huà"]],
+"打开天窗说亮话": [["dǎ"], ["kāi"], ["tiān"], ["chuāng"], ["shuō"], ["liàng"], ["huà"]],
 "打擂台": [["dǎ"], ["lèi"], ["tái"]],
 "打冷枪": [["dǎ"], ["lěng"], ["qiāng"]],
 "打冷战": [["dǎ"], ["lěng"], ["zhàn"]],
@@ -27066,7 +27072,7 @@ module.exports = {
 "金科玉条": [["jīn"], ["kē"], ["yù"], ["tiáo"]],
 "金匮石室": [["jīn"], ["guì"], ["shí"], ["shì"]],
 "金兰之契": [["jīn"], ["lán"], ["zhī"], ["qì"]],
-"金苹果": [["jīn"], ["pín"], ["guǒ"]],
+"金苹果": [["jīn"], ["píng"], ["guǒ"]],
 "金瓶落井": [["jīn"], ["píng"], ["luò"], ["jǐng"]],
 "金融寡头": [["jīn"], ["róng"], ["guǎ"], ["tóu"]],
 "金沙江": [["jīn"], ["shā"], ["jiāng"]],
@@ -28823,6 +28829,7 @@ module.exports = {
 "流行坎止": [["liú"], ["xíng"], ["kǎn"], ["zhǐ"]],
 "流行歌曲": [["liú"], ["xíng"], ["gē"], ["qǔ"]],
 "流行性感冒": [["liú"], ["xíng"], ["xìng"], ["gǎn"], ["mào"]],
+"流行音乐": [["liú"], ["xíng"], ["yīn"], ["yuè"]],
 "流血漂卤": [["liú"], ["xuè"], ["piāo"], ["lǔ"]],
 "流血千里": [["liú"], ["xuè"], ["qiān"], ["lǐ"]],
 "流言飞语": [["liú"], ["yán"], ["fēi"], ["yǔ"]],
@@ -36885,6 +36892,7 @@ module.exports = {
 "行不苟合": [["xíng"], ["bù"], ["gǒu"], ["hé"]],
 "行不履危": [["xíng"], ["bù"], ["lǚ"], ["wēi"]],
 "行不通": [["xíng"], ["bù"], ["tōng"]],
+"行不行": [["xíng"], ["bù"], ["xíng"]],
 "行不胜衣": [["xíng"], ["bù"], ["shèng"], ["yī"]],
 "行步如飞": [["xíng"], ["bù"], ["rú"], ["fēi"]],
 "行不逾方": [["xíng"], ["bù"], ["yú"], ["fāng"]],
@@ -37136,6 +37144,7 @@ module.exports = {
 "学识渊博": [["xué"], ["shí"], ["yuān"], ["bó"]],
 "学疏才浅": [["xué"], ["shū"], ["cái"], ["qiǎn"]],
 "学术界": [["xué"], ["shù"], ["jiè"]],
+"雪茄": [["xuě"], ["jiā"]],
 "雪北香南": [["xuě"], ["běi"], ["xiāng"], ["nán"]],
 "雪耻报仇": [["xuě"], ["chǐ"], ["bào"], ["chóu"]],
 "雪窗萤几": [["xuě"], ["chuāng"], ["yíng"], ["jǐ"]],
@@ -38513,6 +38522,10 @@ module.exports = {
 "音节文字": [["yīn"], ["jié"], ["wén"], ["zì"]],
 "音容宛在": [["yīn"], ["róng"], ["wǎn"], ["zài"]],
 "音信杳无": [["yīn"], ["xìn"], ["yǎo"], ["wú"]],
+"音乐剧": [["yīn"], ["yuè"], ["jù"]],
+"音乐老师": [["yīn"], ["yuè"], ["lǎo"], ["shī"]],
+"音乐喷泉": [["yīn"], ["yuè"], ["pēn"], ["quán"]],
+"音乐频道": [["yīn"], ["yuè"], ["pín"], ["dào"]],
 "殷鉴不远": [["yīn"], ["jiàn"], ["bù"], ["yuǎn"]],
 "殷浩书空": [["yīn"], ["hào"], ["shū"], ["kōng"]],
 "殷天蔽日": [["yīn"], ["tiān"], ["bì"], ["rì"]],
@@ -39062,6 +39075,8 @@ module.exports = {
 "鱼游濠上": [["yú"], ["yóu"], ["háo"], ["shàng"]],
 "鱼游燋釜": [["yú"], ["yóu"], ["zhuó"], ["fǔ"]],
 "娱乐场": [["yú"], ["lè"], ["chǎng"]],
+"娱乐圈": [["yú"], ["lè"], ["quān"]],
+"娱乐频道": [["yú"], ["lè"], ["pín"], ["dào"]],
 "鱼与熊掌": [["yú"], ["yǔ"], ["xióng"], ["zhǎng"]],
 "渔夺侵牟": [["yú"], ["duó"], ["qīn"], ["móu"]],
 "娱妻弄子": [["yú"], ["qī"], ["nòng"], ["zǐ"]],
@@ -40277,6 +40292,7 @@ module.exports = {
 "中华民族": [["zhōng"], ["huá"], ["mín"], ["zú"]],
 "中华民国": [["zhōng"], ["huá"], ["mín"], ["guó"]],
 "中华人民共和国": [["zhōng"], ["huá"], ["rén"], ["mín"], ["gòng"], ["hé"], ["guó"]],
+"中奖": [["zhòng"], ["jiǎng"]],
 "中馈乏人": [["zhōng"], ["kuì"], ["fá"], ["rén"]],
 "中立不倚": [["zhōng"], ["lì"], ["bù"], ["yǐ"]],
 "中馈犹虚": [["zhōng"], ["kuì"], ["yóu"], ["xū"]],
@@ -41127,3 +41143,5 @@ module.exports = {
 "做一日和尚撞一天钟": [["zuò"], ["yī"], ["rì"], ["hé"], ["shàng"], ["zhuàng"], ["yī"], ["tiān"], ["zhōng"]],
 "做一天和尚撞一天钟": [["zuò"], ["yī"], ["tiān"], ["hé"], ["shàng"], ["zhuàng"], ["yī"], ["tiān"], ["zhōng"]],
 };
+
+export default phrases_dict;
