@@ -1,5 +1,5 @@
 // 带声调字符。
-module.exports = {
+const phonetic_symbol: Record<string,string> = {
   "ā": "a1",
   "á": "a2",
   "ǎ": "a3",
@@ -28,3 +28,5 @@ module.exports = {
   "ň": "n3",
   "": "m2",
 };
+
+export default phonetic_symbol;
